weekly dashboard 를 node-cron 에서 github action으로 마이그레이션
