+++ .github/workflows/weekly-dashboard.yml
@@ -1,0 +1,34 @@
+name: send-weekly-dashboard
+
+on:
+  schedule:
+    - cron: '20 2 * * 1' # 월요일 한국시간 오전 11시 20분
+
+jobs:
+  cron:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        node-version: [18.x]
+
+    steps:
+      - name: Check out code
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 2
+
+      - name: Setup Node.js environment
+        uses: actions/setup-node@v3
+        with:
+          node-version: 18
+          cache: 'yarn'
+
+      # cron job을 실행합니다.
+      - name: run script
+        run: |
+          yarn install
+          SLACK_AUTH_TOKEN=${{ secrets.SLACK_AUTH_TOKEN }} \
+          SLACK_ACTIVE_CHANNEL_ID=${{ secrets.SLACK_ACTIVE_CHANNEL_ID }} \
+          SLACK_WATCHER_CHANNEL_ID=${{ secrets.SLACK_WATCHER_CHANNEL_ID }} \
+          GHP_ACCESS_TOKEN=${{ secrets.GHP_ACCESS_TOKEN }} \
+            yarn send:weekly-summary
