+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "pupuri-bot",
-  "version": "0.0.1",
+  "version": "1.0.0",
   "description": "wafflestudio pupuri bot",
   "main": "index.ts",
   "repository": "https://github.com/wafflestudio/pupuri-bot",
@@ -8,31 +8,29 @@
   "license": "MIT",
   "private": false,
   "scripts": {
-    "start": "node dist/src/index.js",
-    "build": "tsc -p .",
-    "dev": "nodemon --watch \"src/**/*.ts\" --exec \"ts-node\" src/index.ts",
-    "deploy": "scripts/deploy.sh",
+    "build:server": "tsc -p .",
+    "start:server": "node dist/src/index.js",
+    "dev:server": "ts-node src/server.ts",
+    "send:weekly-dashboard": "ts-node src/weekly-dashboard.ts",
+    "deploy-server": "scripts/deploy.sh",
     "test:unit": "jest",
     "lint": "eslint ./src"
   },
   "dependencies": {
-    "dotenv": "16.0.3",
-    "express": "4.18.2",
-    "node-cron": "3.0.2"
+    "express": "4.18.2"
   },
   "devDependencies": {
     "@jest/globals": "29.5.0",
     "@types/express": "4.17.17",
     "@types/node": "18.15.10",
-    "@types/node-cron": "3.0.7",
     "@typescript-eslint/eslint-plugin": "5.57.0",
     "@typescript-eslint/parser": "5.57.0",
+    "dotenv": "16.0.3",
     "eslint": "8.37.0",
     "eslint-config-prettier": "8.8.0",
     "eslint-plugin-prettier": "4.2.1",
     "eslint-plugin-simple-import-sort": "10.0.0",
     "jest": "29.5.0",
-    "nodemon": "2.0.22",
     "prettier": "2.8.7",
     "ts-jest": "29.0.5",
     "ts-node": "10.9.1",
